+++ tests/testthat/test-11-RapUnsolved.R
@@ -423,9 +423,9 @@ test_that("solve.RapUnsolved (manual solution, omitting all attribute space plan
   sim_ru <- spp.subset(sim_ru, 1)
   sim_ru <- dp.subset(sim_ru, species = 1, space = 1, points = 1:10)
   sim_ru@data@pu.species.probabilities <- sim_ru@data@pu.species.probabilities[-1, , drop = FALSE]
-  sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@planning.unit.points@coords <- 
+  sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@planning.unit.points@coords <-
     sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@planning.unit.points@coords[-1, , drop = FALSE]
-  sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@planning.unit.points@ids <- 
+  sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@planning.unit.points@ids <-
     sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@planning.unit.points@ids[-1]
   sim_ru@data@targets[[3]] <- c(0.5, 0.1)
   # solve it
@@ -434,7 +434,30 @@ test_that("solve.RapUnsolved (manual solution, omitting all attribute space plan
   expect_equal(space.held(sim_rs, 1)[[1]], -Inf)
 })
 
-
+test_that("solve.RapUnsolved (single demand point)", {
+  # skip if gurobi not installed
+  if (!is.GurobiInstalled(FALSE))
+    skip("Gurobi not installed on system")
+  # load RapUnsolved object
+  set.seed(500)
+  data(sim_ru)
+  sim_ru <- pu.subset(sim_ru, 1:5)
+  sim_ru <- spp.subset(sim_ru, 1)
+  sim_ru <- dp.subset(sim_ru, species = 1, space = 1, points = 1)
+  sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@
+    planning.unit.points@coords[] <- 1
+  sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@
+    demand.points@coords[] <- 1
+  sim_ru@data@attribute.spaces[[1]]@spaces[[1]]@
+    demand.points@weights[] <- 0
+  sim_ru@data@targets[[3]] <- c(0.1, 0)
+  sim_ru@data@pu$cost <- c(5, 1, seq_len(3))
+  # solve it
+  sim_rs <- suppressWarnings(raptr::solve(sim_ru, GurobiOpts(MIPGap = 0)))
+  expect_is(sim_rs, "RapSolved")
+  expect_true(is.na(space.held(sim_rs, 1)[[1]]))
+  expect_equal(selections(sim_rs), c(0, 1, 0, 0, 0))
+})
 
 test_that("update.RapUnsolved", {
   # load RapUnsolved object